requirements.txt
Django
django-classy-tags
